# archive/federation-abandoned-20260206/provisioning-orchestrator/openapi.yaml
openapi: 3.0.3
info:
  title: Federation Platform Provisioning API
  version: 1.0.0
  description: Multi-tenant AIO voice assistant provisioning orchestrator
  contact:
    name: SYNRG Scaling
    email: admin@synrgscaling.com

servers:
  - url: https://federation.synrg.io/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication

  schemas:
    ProvisioningRequest:
      type: object
      required:
        - departmentId
        - departmentName
        - workflows
        - dataRetention
        - region
        - adminEmail
      properties:
        departmentId:
          type: string
          pattern: '^[a-z0-9_-]+$'
          example: 'hr'
          description: Unique department identifier (slug)
        departmentName:
          type: string
          minLength: 1
          maxLength: 255
          example: 'Human Resources'
          description: Human-readable department name
        businessType:
          type: string
          example: 'enterprise'
          description: Type of business for this department
        workflows:
          type: array
          items:
            type: string
          minItems: 1
          example: ['email', 'google_drive', 'database']
          description: List of workflow types to deploy
        dataRetention:
          type: string
          example: '90d'
          description: Data retention policy
        region:
          type: string
          example: 'us-west'
          description: Deployment region
        resources:
          type: object
          properties:
            cpu:
              type: string
              example: '1'
            memory:
              type: string
              example: '2Gi'
        adminEmail:
          type: string
          format: email
          example: 'admin@synrgscaling.com'
          description: Admin email for this department
        googleWorkspaceDomain:
          type: string
          example: 'synrgscaling.com'
          description: Google Workspace domain
        enabledTools:
          type: array
          items:
            type: string
            enum: [email, google_drive, database, vector_store, agent_context, file_download_email]
          example: ['email', 'google_drive', 'database']
        customConfig:
          type: object
          additionalProperties: true
          description: Department-specific custom configuration

    ProvisioningResponse:
      type: object
      properties:
        provisioningId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
          description: Unique provisioning job ID
        departmentId:
          type: string
          example: 'hr'
        status:
          type: string
          enum: [pending, validating, creating_database, creating_oauth, deploying_container, deploying_workflows, configuring_gateway, validating_deployment, completed, failed, rolling_back]
          example: 'completed'
        message:
          type: string
          example: 'Department provisioning completed'
        statusUrl:
          type: string
          example: '/api/provision-status/550e8400-e29b-41d4-a716-446655440000'
        deploymentUrl:
          type: string
          example: 'https://aio-hr.railway.app'
        dashboardUrl:
          type: string
          example: 'https://federation.synrg.io/dashboard/hr'
        errorMessage:
          type: string
          example: 'Database creation failed'

    ProvisioningStatus:
      type: object
      properties:
        provisioningId:
          type: string
          format: uuid
        departmentId:
          type: string
        status:
          type: string
          enum: [pending, validating, creating_database, creating_oauth, deploying_container, deploying_workflows, configuring_gateway, validating_deployment, completed, failed, rolling_back]
        progress:
          type: object
          properties:
            percentage:
              type: integer
              minimum: 0
              maximum: 100
              example: 75
            database:
              type: string
              enum: [pending, in_progress, completed, failed]
            railway:
              type: string
              enum: [pending, in_progress, completed, failed]
            docker:
              type: string
              enum: [pending, in_progress, completed, failed]
            n8n:
              type: string
              enum: [pending, in_progress, completed, failed]
            oauth:
              type: string
              enum: [pending, in_progress, completed, failed]
        currentStep:
          type: string
          example: 'Deploying container to Railway'
        startedAt:
          type: string
          format: date-time
        estimatedCompletion:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        logs:
          type: array
          items:
            type: string
          example:
            - '[2026-02-06T10:00:00Z] [INFO] Provisioning started for department: Human Resources'
            - '[2026-02-06T10:00:05Z] [INFO] Configuration validated successfully'
        errorMessage:
          type: string

    DeprovisioningResponse:
      type: object
      properties:
        departmentId:
          type: string
          example: 'hr'
        status:
          type: string
          enum: [DEPROVISIONING]
          example: 'DEPROVISIONING'
        message:
          type: string
          example: 'Department deprovisioning initiated'
        cleanupProgress:
          type: integer
          minimum: 0
          maximum: 100
          example: 0

    DepartmentList:
      type: object
      properties:
        departments:
          type: array
          items:
            $ref: '#/components/schemas/Department'
        total:
          type: integer
          example: 3

    Department:
      type: object
      properties:
        id:
          type: string
          format: uuid
        departmentId:
          type: string
          example: 'hr'
        departmentName:
          type: string
          example: 'Human Resources'
        status:
          type: string
          enum: [PROVISIONING, ACTIVE, DEPROVISIONING, FAILED, DEPROVISIONED]
        tenantSchema:
          type: string
          example: 'hr_tenant'
        railwayProjectId:
          type: string
        railwayDeploymentUrl:
          type: string
          example: 'https://aio-hr.railway.app'
        enabledTools:
          type: array
          items:
            type: string
        adminEmail:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    HealthCheckResponse:
      type: object
      properties:
        departmentId:
          type: string
          example: 'hr'
        healthy:
          type: boolean
          example: true
        checks:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: 'railway_agent'
              healthy:
                type: boolean
                example: true
              latency:
                type: integer
                example: 45
                description: Latency in milliseconds
              error:
                type: string
        lastCheck:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string
          example: 'Validation error'
        message:
          type: string
          example: 'Department ID is required'
        details:
          type: array
          items:
            type: object

paths:
  /provision-department:
    post:
      summary: Provision a new department AIO instance
      description: Initiates provisioning of a complete department infrastructure including database, OAuth, container, workflows, and gateway configuration
      operationId: provisionDepartment
      tags:
        - Provisioning
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisioningRequest'
      responses:
        '202':
          description: Provisioning initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvisioningResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /provision-status/{provisioningId}:
    get:
      summary: Get provisioning status
      description: Retrieves the current status and progress of a provisioning job
      operationId: getProvisioningStatus
      tags:
        - Provisioning
      security:
        - BearerAuth: []
      parameters:
        - name: provisioningId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Provisioning job ID
      responses:
        '200':
          description: Provisioning status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvisioningStatus'
        '404':
          description: Provisioning job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /deprovision/{departmentId}:
    delete:
      summary: Deprovision a department AIO instance
      description: Initiates cleanup and removal of all department resources
      operationId: deprovisionDepartment
      tags:
        - Provisioning
      security:
        - BearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
          description: Department ID
      responses:
        '202':
          description: Deprovisioning initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeprovisioningResponse'
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /departments:
    get:
      summary: List all departments
      description: Retrieves a list of all provisioned departments
      operationId: listDepartments
      tags:
        - Departments
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of departments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentList'

  /departments/{departmentId}/health:
    get:
      summary: Get department health status
      description: Runs health checks on all department components
      operationId: getDepartmentHealth
      tags:
        - Departments
      security:
        - BearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
          description: Department ID
      responses:
        '200':
          description: Health check results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

tags:
  - name: Provisioning
    description: Department provisioning and deprovisioning operations
  - name: Departments
    description: Department management and health monitoring